requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
requests==2.32.5
numpy==1.26.4
onnxruntime==1.20.1
pyyaml==6.0.2
